requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
distlib==0.3.5
Django==4.1
djangorestframework==3.13.1
filelock==3.7.1
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
virtualenv==20.16.3
